+++ tests/testthat/test-retry.R
@@ -16,4 +16,6 @@ test_that("Check if retry_api is working as intended",{
   
   # this function will fail twice before succeeding
   expect_true(retry_api_call(fail_function(fail_env$i , 3)))
+  expect_error(noctua_options(retry = - 10))
+  expect_error(noctua_options(retry_quiet = "blah"))
 })